requirements.txt
django>=3.1,<3.2
djangorestframework>=3.12,<3.13
django-countries>=7.2.1,< 7.3
django-phonenumber-field>=5.2.0,<5.3
pyyaml>=5.4.1,<5.5
uritemplate>=3.0.1,<3.1
flake8>=3.9.2,<3.10
flake8-isort>=4.0.0,<4.1
psycopg2-binary>=2.9.1,<2.10